% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communication.R
\name{CondComScores}
\alias{CondComScores}
\title{CondComScores: calculate communication scores for each condition in a Seurat Object}
\usage{
ComScores(  
  Seurat,
  idents = "cell_type",
  condition = "orig.ident",
  database = database,
  threshold = 0
)
}
\arguments{
\item{Seurat}{a Seurat object}
\item{idents}{the name of the Seurat column metadata that contains the idents for which the communication will be calculated.}
\item{conditions}{string, specify column in seurat meta.data containing the conditions}
\item{database}{a dataframe with two columns containing ligand and receptors, default is the cellchat database.}
\item{threshold}{minimum percentage of cells expressing the cells in source and target populations. Default is set to 0}
}
\value{
A dataframe including communication scores for each LR pair and interaction
}
\description{
CondComScores: calculate communication scores for each condition in a Seurat Object
}
