% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communication.R
\name{DifComScores}
\alias{DifComScores}
\title{DifComScores: calculate communication scores for each condition in a Seurat Object}
\usage{
DifComScores(seurat, idents = levels(seurat), conditions, cell_type)
}
\arguments{
\item{idents}{a vector containing the list of idents to calculate communication scores, default is levels(Seurat)}

\item{conditions}{string, specify column in seurat meta.data containing the conditions}

\item{cell_type}{string, specify column in seurat meta.data containing the idents}

\item{Seurat}{a Seurat object}
}
\value{
A list of dataframes including communication scores for each LR pair and interaction
}
\description{
DifComScores: calculate communication scores for each condition in a Seurat Object
}
