% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudospatial.R
\name{DPGEA}
\alias{DPGEA}
\title{DPGEA: differential pseudospatial gradient expression analysis}
\usage{
DPGEA(cds, cell_type, condition)
}
\arguments{
\item{cds}{a cds file where pseudotime has already been calculated}

\item{cell_type}{the column of "cds@clusters@listData[["UMAP"]]" that contains cluster names}

\item{condition}{the column of "cds@clusters@listData that contains the condition names}
}
\value{
A list of dataframes including correlations, p-values and the final classification
}
\description{
DPGEA: differential pseudospatial gradient expression analysis
}
