% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communication.R
\name{ComScores}
\alias{ComScores}
\title{ComScores: calculate communication scores}
\usage{
ComScores(Seurat, idents = levels(Seurat))
}
\arguments{
\item{Seurat}{a Seurat object}

\item{idents}{a vector containing the list of idents to calculate communication scores, default is levels(Seurat)}
}
\value{
A list of dataframes including communication scores for each LR pair and interaction
}
\description{
ComScores: calculate communication scores
}
