% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Communication.R
\name{DifComScores}
\alias{DifComScores}
\title{DifComScores: perform differential communication analysis between each condition}
\usage{
DifComScores(
  seurat,
  idents = "cell_type",
  condition = "orig.ident",
  CondComScores,
  order = c("fed_int","fast12_int","fast24_int")
)
}
\arguments{

\item{Seurat}{ a Seurat object}
\item{idents}{ a vector containing the list of idents to calculate communication scores, default is levels(Seurat)}
\item{conditions}{ string, specify column in seurat meta.data containing the conditions}
\item{CondComScores}{ the output of the CondComScores function}
\item{order}{ a vector, order of the conditions to be compared for the contrast}

}
\value{
A dataframe 
}
\description{
DifComScores: perform differential communication analysis between each condition 
}
